Streamlined lsm303dlhc driver
