+++ sw/airborne/peripherals/lsm303dlhc_i2c.h
@@ -24,6 +24,7 @@
  * @file peripherals/lsm303dlhc.h
  *
  * Driver for ST LSM303DLHC 3D accelerometer and magnetometer.
+ * UNTESTED
  */
 
 #ifndef LSM303DLHC_H
@@ -33,41 +34,11 @@
 #include "mcu_periph/i2c.h"
 #include "math/pprz_algebra_int.h"
 
-/* Address and register definitions */
-#include "peripherals/lsm303dlhc_regs.h"
+#include "peripherals/lsm303dlhc.h"
 
-struct Lsm303dlhcAccConfig {
-  uint8_t rate;    ///< Data Output Rate Bits(6 -> 50Hz with HMC5843, 75Hz with HMC5883)
-  uint8_t lp_mode; ///< Low power mode
-  uint8_t scale;   ///< full scale selection
-  uint8_t hres;    ///< high resolution output mode
-};
 
-struct Lsm303dlhcMagConfig {
-  uint8_t rate;  ///< Data Output Rate Bits(6 -> 50Hz with HMC5843, 75Hz with HMC5883)
-  uint8_t gain;  ///< Gain configuration (1 -> +- 1 Gauss)
-  uint8_t mode;  ///< Measurement mode
-};
 
-/** config status states */
-enum Lsm303dlhcAccConfStatus {
-  LSM_CONF_ACC_UNINIT,
-  LSM_CONF_ACC_CTRL_REG4_A,
-  LSM_CONF_ACC_CTRL_REG1_A,
-  LSM_CONF_ACC_CTRL_REG3_A,
-  LSM_CONF_ACC_DONE
-};
-
-/** config status states */
-enum Lsm303dlhcMagConfStatus {
-  LSM_CONF_MAG_UNINIT,
-  LSM_CONF_MAG_CRA_REG_M,
-  LSM_CONF_MAG_CRB_REG_M,
-  LSM_CONF_MAG_MR_REG_M,
-  LSM_CONF_MAG_DONE
-};
-
-struct Lsm303dlhc {
+struct Lsm303dlhc_i2c {
   struct i2c_periph *i2c_p;
   struct i2c_transaction i2c_trans;
   bool initialized;                 ///< config done flag
@@ -85,8 +56,6 @@ struct Lsm303dlhc {
     struct Lsm303dlhcMagConfig mag;
   } config;
 };
-
-
 
 // TODO IRQ handling
